[misc] Minor API rework for consistency with Mujoco API. (#283)
* [misc] Move to C++17 for dependencies.
* [core] Compute subtree coms by default.
* [core] Improve vector 'clamp' utility to avoid memory allocation.
* [core] Fix edge case in controller breakpoints management.
* [core] Rename 'rotorInertia' in 'armature' to be consistent with mujoco and clearer that it is joint side.
* [core] Rename 'EffortLimit' in 'ControlLimit' to avoid redundancy with pinocchio::Data and explicit it is motor-side.
* [core] Adapt code base for new 'armuture' and 'controlLimit' definitions.
* [core] Do not compute 'oMf' for 'JOINT' and 'BODY' frames but rather copy data from 'oMi' and 'oMf'.
* [core] Return static variables for Robot'getControlLimit' and 'getArmatures' to avoid memory allocations.
* [core] Improvement breakpoint management to avoid skipping controller/sensor updates.
* [core|python] Add option to monitor motor efforts and plot them by default instead of command.
* [core|python/viewer] Provide generic interpolation method for configuration evolution over time. Use it in Python viewer.
* [core|python|gym] Add dedicated 'u_custom' field for controller's internal dynamics. Rename 'u_command' in 'command' to clarify it does not have to be an effort. 
* [core] Split evaluation of custom internal dynamics from position/velocity bounds and flexibility dynamics. Replace 'u' with more specific variable name. 
* [core] Set 'command' and 'u_custom' to zero automatically before calling callbacks. 
* [core] Rename 'ControlLimit' in 'CommandLimit' for consistency.
* [core/python] Make sure registered variables vector is 1D.
* [python/robot] Enable setting options only partially.
* [python/dynamics] Add optional flag to update collisions. Fix com jacobian not computed even if requested.
* [python/simulator] Do not forbid to render system at if no simulation is running.
* [gym/common] Ensure float64 action buffer for compatibility with register_variable telemetry.
* [gym/common/envs] Increment 'num_steps' only at the end of successful step, not right before integration.
* [gym/common/envs] Set robot in neutral configuration at init.
* [gym/common/envs] Fix edge case for robot without sensors.
* [gym/common/envs] Do NOT enforce position/velocity/effort bounds at _setup to avoid inconsistent observation/action spaces bounds.

+++ python/gym_jiminy/common/gym_jiminy/common/wrappers/frame_stack.py
@@ -161,7 +161,7 @@ class StackedJiminyEnv(BasePipelineWrapper):
 
         # Initialize some internal buffers
         self.__n_last_stack = 0
-        self._action = zeros(self.action_space)
+        self._action = zeros(self.action_space, dtype=np.float64)
         self._observation = zeros(self.observation_space)
 
     def _setup(self) -> None:
